Rewrite backup job to use a cronjob instead of a tekton task
With the recent changes [1] to how we are handling vault configuration, we
now have two mechanisms for triggering tasks: one using `oc create job`,
and another one involving tekton tasks and listeners.

This commit rewrites the backup job to use a CronJob instead of a tekton
Task. In place of the task steps, we take advantage of the fact that
initContainers are executed sequentially before the main container starts.

[1]: https://github.com/OCP-on-NERC/nerc-ocp-config/pull/517