Enable wildcards in routeAdmission policy
This is necessary to support kubevirt clusters that don't have an explicit
base domain [1]. The ingress for these clusters is a child of the managed
cluster ingress.

For example, if the managed cluster ingress is *.apps.cluster.example.com,
then a kubevirt cluster without an explicit base domain will use the
ingress *.apps.guest.apps.cluster.example.com.

[1]: https://docs.openshift.com/container-platform/4.18/hosted_control_planes/hcp-deploy/hcp-deploy-virt.html#hcp-virt-ingress-dns_hcp-deploy-virt